Changed enum values as per opcode

+++ src/main/java/com/zoltanbalazs/Access_Flags.java
@@ -10,19 +10,19 @@ public interface Access_Flags {
 }
 
 enum Class_Access_Flags implements Access_Flags {
-    ACC_PUBLIC((short) 0x0001),
-    ACC_FINAL((short) 0x0010),
-    ACC_SUPER((short) 0x0020),
-    ACC_INTERFACE((short) 0x0200),
-    ACC_ABSTRACT((short) 0x0400),
-    ACC_SYNTHETIC((short) 0x1000),
-    ACC_ANNOTATION((short) 0x2000),
-    ACC_ENUM((short) 0x4000);
+    ACC_PUBLIC(0x0001),
+    ACC_FINAL(0x0010),
+    ACC_SUPER(0x0020),
+    ACC_INTERFACE(0x0200),
+    ACC_ABSTRACT(0x0400),
+    ACC_SYNTHETIC(0x1000),
+    ACC_ANNOTATION(0x2000),
+    ACC_ENUM(0x4000);
 
     public final short value;
 
-    private Class_Access_Flags(short value) {
-        this.value = value;
+    private Class_Access_Flags(int value) {
+        this.value = (short) value;
     }
 
     public static List<Class_Access_Flags> parseFlags(short value) {
@@ -39,20 +39,20 @@ enum Class_Access_Flags implements Access_Flags {
 }
 
 enum Field_Access_Flags implements Access_Flags {
-    ACC_PUBLIC((short) 0x0001),
-    ACC_PRIVATE((short) 0x0002),
-    ACC_PROTECTED((short) 0x0004),
-    ACC_STATIC((short) 0x0008),
-    ACC_FINAL((short) 0x0010),
-    ACC_VOLATILE((short) 0x0040),
-    ACC_TRANSIENT((short) 0x0080),
-    ACC_SYNTHETIC((short) 0x1000),
-    ACC_ENUM((short) 0x4000);
+    ACC_PUBLIC(0x0001),
+    ACC_PRIVATE(0x0002),
+    ACC_PROTECTED(0x0004),
+    ACC_STATIC(0x0008),
+    ACC_FINAL(0x0010),
+    ACC_VOLATILE(0x0040),
+    ACC_TRANSIENT(0x0080),
+    ACC_SYNTHETIC(0x1000),
+    ACC_ENUM(0x4000);
 
     public final short value;
 
-    private Field_Access_Flags(short value) {
-        this.value = value;
+    private Field_Access_Flags(int value) {
+        this.value = (short) value;
     }
 
     public static List<Field_Access_Flags> parseFlags(short value) {
@@ -69,23 +69,23 @@ enum Field_Access_Flags implements Access_Flags {
 }
 
 enum Method_Access_Flags implements Access_Flags {
-    ACC_PUBLIC((short) 0x0001),
-    ACC_PRIVATE((short) 0x0002),
-    ACC_PROTECTED((short) 0x0004),
-    ACC_STATIC((short) 0x0008),
-    ACC_FINAL((short) 0x0010),
-    ACC_SYNCHRONIZED((short) 0x0020),
-    ACC_BRIDGE((short) 0x0040),
-    ACC_VARARGS((short) 0x0080),
-    ACC_NATIVE((short) 0x0100),
-    ACC_ABSTRACT((short) 0x0400),
-    ACC_STRICT((short) 0x0800),
-    ACC_SYNTHETIC((short) 0x1000);
+    ACC_PUBLIC(0x0001),
+    ACC_PRIVATE(0x0002),
+    ACC_PROTECTED(0x0004),
+    ACC_STATIC(0x0008),
+    ACC_FINAL(0x0010),
+    ACC_SYNCHRONIZED(0x0020),
+    ACC_BRIDGE(0x0040),
+    ACC_VARARGS(0x0080),
+    ACC_NATIVE(0x0100),
+    ACC_ABSTRACT(0x0400),
+    ACC_STRICT(0x0800),
+    ACC_SYNTHETIC(0x1000);
 
     public final short value;
 
-    private Method_Access_Flags(short value) {
-        this.value = value;
+    private Method_Access_Flags(int value) {
+        this.value = (short) value;
     }
 
     public static List<Method_Access_Flags> parseFlags(short value) {
